flask-api/requirements.txt
Flask==3.0.3
flask-cors
mysql-connector-python